backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.12.5
pydantic-settings==2.7.1

# Environment
python-dotenv==1.0.1

# LLM APIs
openai==1.58.1
google-generativeai==0.8.3

# Database
supabase==2.10.0

# HTTP Client (compatible with both supabase and openai)
httpx>=0.26,<0.28

# Utilities
numpy==1.26.4
